GhettOSD: fix sats number/altitude overflow.

+++ GhettOSD/GhettOSD/OSD_Vars.h
@@ -114,7 +114,7 @@ static int16_t      off_course;
 static uint8_t      osd_got_home = 0;               // tels if got home position or not
 static float        osd_home_lat = 0;               // home latidude
 static float        osd_home_lon = 0;               // home longitude
-static int          osd_home_alt = 0; 
+static float        osd_home_alt = 0; 
 static int          osd_alt_to_home = 0; 
 static long         osd_home_distance = 0;          // distance from home
 static uint8_t      osd_home_direction;             // Arrow direction pointing to home (1-16 to CW loop)
@@ -124,7 +124,7 @@ static int16_t      osd_roll = 0;                   // roll from DCM
 static float        osd_heading = 0;                // ground course heading from GPS
 static float        glide = 0;
 
-static int          osd_alt = 0;                    // altitude
+static float        osd_alt = 0;                    // altitude
 static float        osd_airspeed = -1;              // airspeed
 static float        osd_windspeed = 0;
 static float        osd_windspeedz = 0;
